+++ src/DataDefinitionsBundle/Controller/ImportRuleController.php
@@ -8,7 +8,7 @@ declare(strict_types=1);
  * Full copyright and license information is available in
  * LICENSE.md which is distributed with this source code.
  *
- * @copyright  Copyright (c) CORS GmbH (https://www.cors.gmbh) in combination with instride AG (https://instride.ch)
+ * @copyright  Copyright (c) CORS GmbH (https://www.cors.gmbh)
  * @license    DDCL
  */
 
@@ -18,15 +18,14 @@ use Instride\Bundle\DataDefinitionsBundle\Form\Type\ImportRulesImportType;
 use OpenSpout\Common\Entity\Row;
 use OpenSpout\Reader\XLSX\Reader;
 use OpenSpout\Writer\XLSX\Writer;
-// Pimcore 2026: classic GDPR\AdminController gone - kwizda compat shim base
-use Pimcore\Bundle\AdminBundle\Controller\AdminAbstractController as AdminController;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\Form\FormFactoryInterface;
 use Symfony\Component\HttpFoundation\BinaryFileResponse;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 
-final class ImportRuleController extends AdminController
+final class ImportRuleController extends AbstractController
 {
     public function importAction(Request $request, FormFactoryInterface $formFactory): JsonResponse
     {