+++ src/DataDefinitionsBundle/DependencyInjection/Configuration.php
@@ -8,7 +8,7 @@ declare(strict_types=1);
  * Full copyright and license information is available in
  * LICENSE.md which is distributed with this source code.
  *
- * @copyright  Copyright (c) CORS GmbH (https://www.cors.gmbh) in combination with instride AG (https://instride.ch)
+ * @copyright  Copyright (c) CORS GmbH (https://www.cors.gmbh)
  * @license    DDCL
  */
 
@@ -35,7 +35,7 @@ final class Configuration implements ConfigurationInterface
     #[\Override]
     public function getConfigTreeBuilder(): TreeBuilder
     {
-        $treeBuilder = new TreeBuilder('instride_data_definitions');
+        $treeBuilder = new TreeBuilder('data_definitions');
         $rootNode = $treeBuilder->getRootNode();
 
         ConfigurationHelper::addConfigLocationWithWriteTargetNodes($rootNode, [
@@ -134,7 +134,6 @@ final class Configuration implements ConfigurationInterface
             ->end()
         ;
 
-        $this->addPimcoreResourcesSection($rootNode);
         $this->addModelsSection($rootNode);
 
         return $treeBuilder;
@@ -191,30 +190,4 @@ final class Configuration implements ConfigurationInterface
         ;
     }
 
-    private function addPimcoreResourcesSection(ArrayNodeDefinition $node)
-    {
-        $node->children()
-            ->arrayNode('pimcore_admin')
-                ->addDefaultsIfNotSet()
-                ->children()
-                    ->arrayNode('js')
-                        ->useAttributeAsKey('name')
-                        ->prototype('scalar')->end()
-                    ->end()
-                    ->arrayNode('css')
-                        ->useAttributeAsKey('name')
-                        ->prototype('scalar')->end()
-                    ->end()
-                    ->arrayNode('install')
-                        ->addDefaultsIfNotSet()
-                        ->ignoreExtraKeys(false)
-                        ->children()
-                            ->scalarNode('sql')->defaultValue(['@DataDefinitionsBundle/Resources/install/pimcore/sql/data.sql'])->end()
-                        ->end()
-                    ->end()
-                ->end()
-            ->end()
-        ->end()
-        ;
-    }
 }