+++ src/DataDefinitionsBundle/Resources/assets/pimcore-studio/src/index.ts
@@ -1,7 +1,7 @@
 /**
  * Data Definitions Bundle - Pimcore Studio Plugin
  *
- * @copyright  Copyright (c) CORS GmbH (https://www.cors.gmbh) in combination with instride AG (https://www.instride.ch)
+ * @copyright  Copyright (c) CORS GmbH (https://www.cors.gmbh)
  * @license    DDCL
  */