+++ src/DataDefinitionsBundle/Resources/public/studio/2d7897cb-93fb-4697-be98-466e2c3b185a/exposeRemote.js
@@ -1,0 +1,13 @@
+
+      if (window.pluginRemotes === undefined) {
+        window.pluginRemotes = {}
+      }
+
+      if (window.alternativePluginExportPaths === undefined) {
+        window.alternativePluginExportPaths = {}
+      }
+
+      window.pluginRemotes.datadefinitions = "/bundles/datadefinitions/studio/2d7897cb-93fb-4697-be98-466e2c3b185a/static/js/remoteEntry.js"
+
+      
+