+++ tests/DataDefinitionsBundle/Behat/Context/Setup/ImportDefinitionContext.php
@@ -8,8 +8,8 @@
  * For the full copyright and license information, please view the LICENSE.md and gpl-3.0.txt
  * files that are distributed with this source code.
  *
- * @copyright 2024 instride AG (https://instride.ch)
- * @license   https://github.com/instride-ch/DataDefinitions/blob/5.0/gpl-3.0.txt GNU General Public License version 3 (GPLv3)
+ * @copyright  Copyright (c) CORS GmbH (https://www.cors.gmbh)
+ * @license    DDCL
  */
 
 namespace Instride\Bundle\DataDefinitionsBundle\Behat\Context\Setup;